+++ Holovibes/includes/io_files/output_frame_file.hh
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "frame_file.hh"
-#include "compute_descriptor.hh"
+#include "logger.hh"
 
 namespace holovibes::io_files
 {
@@ -18,8 +18,8 @@ class OutputFrameFile : public FrameFile
   public:
     /*! \brief Export the compute settings in the file
      *
-     * \param cd The ComputeDescriptor containing the compute settings
-     * \param record_raw Is the raw record enabled
+     * \param input_fps how fast the input was
+     * \param contiguous How many frames were contiguous in input
      */
     virtual void export_compute_settings(int input_fps, size_t contiguous) = 0;