+++ Holovibes/sources/main.cc
@@ -79,7 +79,10 @@ static int start_gui(holovibes::Holovibes& holovibes, int argc, char** argv, con
     holovibes::gui::MainWindow window;
     window.show();
     splash.finish(&window);
-    holovibes.get_cd().register_observer(window);
+
+    // Set callbacks
+    holovibes::GSH::instance().set_notify_callback([&]() { window.notify(); });
+    holovibes::Holovibes::instance().set_error_callback([&](auto e) { window.notify_error(e); });
 
     if (!filename.empty())
     {