+++ Holovibes/sources/state/global_state_holder.cc
@@ -155,6 +155,22 @@ void GSH::set_fft_shift_enabled(bool value)
     api::pipe_refresh();
 }
 
+void GSH::set_composite_p_h(Span<uint> span, bool notify)
+{
+    composite_cache_.get_hsv_ref()->h.p_max = span.min;
+    composite_cache_.get_hsv_ref()->h.p_max = span.max;
+    if (notify)
+        this->notify();
+}
+
+void GSH::set_rgb_p(Span<int> span, bool notify)
+{
+    composite_cache_.get_rgb_ref()->p_min = span.min;
+    composite_cache_.get_rgb_ref()->p_max = span.max;
+    if (notify)
+        this->notify();
+}
+
 void GSH::set_weight_rgb(int r, int g, int b)
 {
     set_weight_r(r);
@@ -273,7 +289,7 @@ void GSH::update_contrast(WindowKind kind, float min, float max)
     window->contrast_min = min;
     window->contrast_max = max;
 
-    update_view(kind);
+    notify();
 }
 
 std::shared_ptr<View_Window> GSH::get_window(WindowKind kind)